requirements.txt
jupyter==1.0.0
matplotlib==2.1.2
numpy==1.14.0
pandas==0.22.0
scikit-learn==0.19.1
scipy==1.0.0
tensorflow-gpu==1.4.1
